move tests, benchmarks and examples to their own directory
