+++ examples/sphincs+-shake.cpp
